refactor: rename app/group/[groupId] to app/group/[groupSlug]
Renames the 13 files under app/group/[groupId]/** to app/group/[groupSlug]/**
and swaps each wrapper's slug resolution direction: instead of parsing a
numeric groupId and resolving its slug (resolveGroupSlugById), each wrapper
now parses the groupSlug URL param and resolves the numeric id
(resolveGroupIdBySlug), 404ing via notFound() when no group has that slug.
The home wrapper's existing redirect-to-/ check for viewing your own group
now compares against the id resolved from the slug.

Part of #490.
