+++ app/group/[groupSlug]/effort/page.tsx
@@ -1,0 +1,17 @@
+import { notFound } from 'next/navigation';
+import { resolveGroupIdBySlug } from '@/lib/group-slug';
+import PayOffPage from '@/app/(routes)/effort/page';
+
+export default async function CrossGroupEffortPage({
+	params
+}: {
+	params: Promise<{ groupSlug: string }>;
+}) {
+	const { groupSlug } = await params;
+	const viewedGroupId = await resolveGroupIdBySlug(groupSlug);
+	if (viewedGroupId === null) {
+		notFound();
+	}
+	const viewedGroup = { id: viewedGroupId, slug: groupSlug };
+	return <PayOffPage viewedGroup={viewedGroup} />;
+}
