+++ app/group/[groupSlug]/pulli/page.tsx
@@ -1,0 +1,17 @@
+import { notFound } from 'next/navigation';
+import { resolveGroupIdBySlug } from '@/lib/group-slug';
+import PulliPage from '@/app/(routes)/pulli/page';
+
+export default async function CrossGroupPulliPage({
+	params
+}: {
+	params: Promise<{ groupSlug: string }>;
+}) {
+	const { groupSlug } = await params;
+	const viewedGroupId = await resolveGroupIdBySlug(groupSlug);
+	if (viewedGroupId === null) {
+		notFound();
+	}
+	const viewedGroup = { id: viewedGroupId, slug: groupSlug };
+	return <PulliPage viewedGroup={viewedGroup} />;
+}
