+++ app/group/[groupSlug]/resightings/page.tsx
@@ -1,0 +1,17 @@
+import { notFound } from 'next/navigation';
+import { resolveGroupIdBySlug } from '@/lib/group-slug';
+import ResightingsPage from '@/app/(routes)/resightings/page';
+
+export default async function CrossGroupResightingsPage({
+	params
+}: {
+	params: Promise<{ groupSlug: string }>;
+}) {
+	const { groupSlug } = await params;
+	const viewedGroupId = await resolveGroupIdBySlug(groupSlug);
+	if (viewedGroupId === null) {
+		notFound();
+	}
+	const viewedGroup = { id: viewedGroupId, slug: groupSlug };
+	return <ResightingsPage viewedGroup={viewedGroup} />;
+}
