+++ app/group/[groupSlug]/session/_shared.tsx
