+++ app/group/[groupSlug]/species/page.tsx
@@ -1,0 +1,17 @@
+import { notFound } from 'next/navigation';
+import { resolveGroupIdBySlug } from '@/lib/group-slug';
+import AllSpeciesPage from '@/app/(routes)/species/page';
+
+export default async function CrossGroupSpeciesPage({
+	params
+}: {
+	params: Promise<{ groupSlug: string }>;
+}) {
+	const { groupSlug } = await params;
+	const viewedGroupId = await resolveGroupIdBySlug(groupSlug);
+	if (viewedGroupId === null) {
+		notFound();
+	}
+	const viewedGroup = { id: viewedGroupId, slug: groupSlug };
+	return <AllSpeciesPage viewedGroup={viewedGroup} />;
+}
